# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tiny Ledger API
  version: 1.0.0
  summary: Event-sourced ledger — accounts, money movements, balances and audit trail.
  description: |
    The contract precedes the code (spec §5): server interfaces and DTOs are generated from this
    document, so a controller that drifts from it fails compilation.

    Conventions (spec §7, §7.1 records each adoption from the Starling Bank public API):

    - Identifiers are UUIDs named `<entity>Uid` — never `id`, never a database key.
    - Money is one shape everywhere: `{"currency": "GBP", "minorUnits": 10000}`. `minorUnits` is an
      `int64` of pence or cents. No float, no decimal string.
    - Money movement is `PUT` to a client-generated movement UID: the UID is both the dedup key and
      the movement's permanent identity (§6.3). There is no idempotency header.
    - Lists are wrapped in a named key and paginated by an opaque cursor (`links.next`).
    - Errors are RFC 7807 problem details (§6.5), never Starling's `{"errors": [...]}`.

    The two auditor operations (`getEvents`, `listAuditEntries`) exist in the `full` profile only;
    the `standalone` profile answers them `501` (§7).
  contact:
    name: Flávio Oliva
  license:
    name: MIT
    identifier: MIT
servers:
  # Path versioning lives in the paths (§7.1), so these carry an origin and nothing else. They are
  # here because a client generated from this file needs somewhere to send: `url: /` gave Postman a
...
  - url: https://app.localhost
...
security:
  - bearerAuth: []
paths:
  /api/v1/accounts:
    post:
      tags: [accounts]
      operationId: openAccount
      summary: Open an account
      description: |
        Requires `ledger:writer`. The `accountUid` is server-generated (§6.3). Ownership *is* enforced.

        `ledger:writer` is permission to write, **not an unlimited entitlement to create**: a caller
        may self-open at most `ledger.accounts.max-per-owner` accounts, and the next one is a `409`.
        How many a customer may hold is a bank decision, not a caller's. The `standalone` profile
        turns the rule off — it runs as one principal, where a per-owner cap caps everything.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Account opened.
          headers:
            Location:
              description: The URI of the account just opened.
              required: true
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: |
            The caller already holds every account the bank lets them self-open (§6.5). A `403` would
            be wrong — the roles are correct; what conflicts is state the caller can resolve.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: /errors/account-limit-reached
                title: Account limit reached
                status: 409
                detail: You already hold the maximum of 2 accounts.
                traceId: 7c6f2e1a9b4d5c30
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
    get:
      tags: [accounts]
      operationId: listAccounts
      summary: List the caller's own accounts
      description: |
        Scoped to the JWT subject — a caller never sees another principal's accounts (§9.3 N12).
        Requires `ledger:reader`. Ownership *is* enforced.
      responses:
        '200':
          description: The caller's accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
  /api/v1/accounts/{accountUid}:
    parameters:
      - $ref: '#/components/parameters/AccountUid'
    get:
      tags: [accounts]
      operationId: getAccount
      summary: Read account metadata
      responses:
        '200':
          description: The account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
  /api/v1/accounts/{accountUid}/deposits/{depositUid}:
    parameters:
      - $ref: '#/components/parameters/AccountUid'
      - name: depositUid
        in: path
        required: true
        description: Client-generated UUID — the deposit's identity and its dedup key (§6.3).
        example: 1f0a9d3c-8e42-4a55-9b1e-6c7d8e9f0a1b
        schema:
          type: string
          format: uuid
    put:
      tags: [movements]
      operationId: putDeposit
      summary: Deposit money
      description: |
        Idempotent by identity (§6.3): the first write answers `201`, a replay of the same UID with
        the same payload answers `200` with the original result, and the same UID with a different
        payload answers `409`. Lookups are global — reusing a UID against a different account is a
        `409`, not a fresh movement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequest'
      responses:
        '201':
          description: Deposit recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '200':
          description: Replay of an identical deposit — the original result, never re-applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: Currency mismatch with the account (§6.5).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                currencyMismatch:
                  $ref: '#/components/examples/CurrencyMismatch'
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
  /api/v1/accounts/{accountUid}/withdrawals/{withdrawalUid}:
    parameters:
      - $ref: '#/components/parameters/AccountUid'
      - name: withdrawalUid
        in: path
        required: true
        description: Client-generated UUID — the withdrawal's identity and its dedup key (§6.3).
        example: 2c1b8e47-3d95-4f06-a8b2-7e5d4c3b2a19
        schema:
          type: string
          format: uuid
    put:
      tags: [movements]
      operationId: putWithdrawal
      summary: Withdraw money
      description: |
        Idempotent by identity (§6.3). Rejections replay deterministically: retrying a refused
        withdrawal with the same UID returns the original `422`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequest'
      responses:
        '201':
          description: Withdrawal recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '200':
          description: Replay of an identical withdrawal — the original result, never re-applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: |
            The movement was refused by a business invariant (§6.5) — insufficient funds, or a
            currency mismatch with the account. The `type` distinguishes them.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                insufficientFunds:
                  summary: Insufficient funds
                  value:
                    type: /errors/insufficient-funds
                    title: Insufficient funds
                    status: 422
                    detail: The account balance is lower than the requested withdrawal.
                    traceId: 7c6f2e1a9b4d5c30
                currencyMismatch:
                  $ref: '#/components/examples/CurrencyMismatch'
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
  /api/v1/accounts/{accountUid}/asset-transfers/{transferUid}:
    parameters:
      - $ref: '#/components/parameters/AccountUid'
      - name: transferUid
        in: path
        required: true
        description: Client-generated UUID — the asset transfer's identity and its dedup key (§6.3).
        example: 3b2a19c7-4e82-4f06-b9d3-8e7d5c4b1a20
        schema:
          type: string
          format: uuid
    put:
      tags: [movements]
      operationId: putAssetTransfer
      summary: Transfer ETF assets with tax-lot allocation
      description: |
        Idempotent by identity (§6.3). Supports inbound ETF asset transfer (acquisition with cost-basis and lot ID)
        and outbound ETF asset transfer (disposal with tax-lot selection such as HIFO or FIFO).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetTransferRequest'
      responses:
        '201':
          description: Asset transfer recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetTransaction'
        '200':
          description: Replay of an identical asset transfer — the original result, never re-applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetTransaction'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: |
            The movement was refused by a business invariant (§6.5) — insufficient asset holding, asset mismatch, or
            cost basis currency mismatch.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                insufficientHolding:
                  summary: Insufficient holding
                  value:
                    type: /errors/insufficient-holding
                    title: Insufficient holding
                    status: 422
                    detail: The account asset holding is lower than the requested transfer.
                    traceId: 7c6f2e1a9b4d5c30
                currencyMismatch:
                  $ref: '#/components/examples/CurrencyMismatch'
                assetMismatch:
                  summary: Asset mismatch
                  value:
                    type: /errors/asset-mismatch
                    title: Asset mismatch
                    status: 422
                    detail: Asset class or symbol mismatch.
                    traceId: 7c6f2e1a9b4d5c30
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
  /api/v1/accounts/{accountUid}/balance:
    parameters:
      - $ref: '#/components/parameters/AccountUid'
    get:
      tags: [balance]
      operationId: getBalance
      summary: Read the account balance
      description: |
        Served from the projection by default, which is eventually consistent — `asOf` and
        `streamVersion` are the staleness markers (§4.4). `consistency=strong` bypasses cache and
        projection, folding the event stream instead.
      parameters:
        - name: consistency
          in: query
          required: false
          description: |
            `strong` reads through to the event store. Omit for the projection read.
          example: strong
          schema:
            type: string
            enum: [strong]
      responses:
        '200':
          description: The balance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/EventStoreUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      description: |
        **The API requires only a valid JWT** — it is a resource server and does not care how the
        token was obtained (§6.4). The flow below is declared so a generated client can fetch one
        without being told out of band, and it names this repository's **dev fixture**: Direct
        Access Grants against the Keycloak realm the Compose stack starts. A real deployment
        replaces `tokenUrl` and the client. The `standalone` profile needs none of this — it is
        unauthenticated and runs with a fixed local principal (§1).
      flows:
        password:
          tokenUrl: https://auth.localhost/realms/tiny-ledger/protocol/openid-connect/token
          refreshUrl: https://auth.localhost/realms/tiny-ledger/protocol/openid-connect/token
          scopes: {}
  parameters:
    AccountUid:
      name: accountUid
      in: path
      required: true
      description: |
        The account's unguessable UUID. Server-generated, so the example is illustrative — use the
        one `openAccount` returned.
      example: f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: |
        Malformed shape (§6.5) — zero, negative or non-integer `minorUnits`, a bad currency code, an
        unreadable body, or a query parameter outside its declared range or enum. `type` is
        `/errors/invalid-amount`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/invalid-amount
            title: Invalid amount
            status: 400
            detail: The request is malformed. minorUnits must be a positive integer and currency a valid ISO-4217 code.
            traceId: 7c6f2e1a9b4d5c30
    Unauthenticated:
      description: No credential, or one that does not validate. `type` is `/errors/unauthenticated`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/unauthenticated
            title: Unauthenticated
            status: 401
            detail: No valid credential was presented.
            traceId: 7c6f2e1a9b4d5c30
    Forbidden:
      description: |
        Wrong role or wrong owner (§6.5). Wrong-owner access returns `403`, not `404` — the
        account-existence oracle is accepted because `accountUid`s are unguessable UUIDs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/forbidden
            title: Forbidden
            status: 403
            detail: You do not have access to this account.
            traceId: 7c6f2e1a9b4d5c30
    AccountNotFound:
      description: No such account. `type` is `/errors/account-not-found`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/account-not-found
            title: Account not found
            status: 404
            detail: No such account.
            traceId: 7c6f2e1a9b4d5c30
    Conflict:
      description: |
        Either the movement UID was reused with a different payload
        (`/errors/idempotency-conflict` — the original movement stands), or the append lost an
        optimistic-concurrency race (`/errors/version-conflict`). The `type` distinguishes them.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            idempotencyConflict:
              summary: Reused movement UID, different payload
              value:
                type: /errors/idempotency-conflict
                title: Idempotency conflict
                status: 409
                detail: This movement UID already identifies a different movement.
                traceId: 7c6f2e1a9b4d5c30
            versionConflict:
              summary: Concurrent modification
              value:
                type: /errors/version-conflict
                title: Version conflict
                status: 409
                detail: The account stream moved on. Retry the command.
                traceId: 7c6f2e1a9b4d5c30
    RateLimited:
      description: Rate limit exceeded (§6.1). `type` is `/errors/rate-limit-exceeded`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          required: true
          schema:
            type: integer
            format: int32
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/rate-limit-exceeded
            title: Rate limit exceeded
            status: 429
            detail: Too many requests. Retry after the interval in the Retry-After header.
            traceId: 7c6f2e1a9b4d5c30
...
    EventStoreUnavailable:
      description: The event store is unreachable. `type` is `/errors/event-store-unavailable`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          required: true
          schema:
            type: integer
            format: int32
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: /errors/event-store-unavailable
            title: Event store unavailable
            status: 503
            detail: The event store is unreachable. Retry after the interval in the Retry-After header.
            traceId: 7c6f2e1a9b4d5c30
  examples:
    CurrencyMismatch:
      summary: Currency mismatch with the account
      value:
        type: /errors/currency-mismatch
        title: Currency mismatch
        status: 422
        detail: The amount currency differs from the account currency.
        traceId: 7c6f2e1a9b4d5c30
  schemas:
...
    MovementAmount:
      type: object
      description: |
        `Money` as a movement may carry it: strictly positive. Zero, negative or non-integer
        `minorUnits` is a `400` `/errors/invalid-amount` (§6.5); the sign lives in `direction`, not
        in the amount.
      required: [currency, minorUnits]
      properties:
        currency:
          type: string
          description: ISO-4217 alphabetic code. Must match the account's, or the write is a `422`.
          pattern: '^[A-Z]{3}$'
          example: GBP
        minorUnits:
          type: integer
          format: int64
          minimum: 1
          example: 10000
    OpenAccountRequest:
      type: object
      required: [name, currency]
      properties:
        name:
          type: string
          description: The account's human-readable name — the CLI resolves names to UIDs with it.
          minLength: 1
          maxLength: 140
          example: ACC-001
        currency:
          type: string
          description: ISO-4217 alphabetic code. Fixed for the life of the account.
          pattern: '^[A-Z]{3}$'
          example: GBP
    MovementRequest:
      type: object
      required: [amount]
      properties:
        amount:
          $ref: '#/components/schemas/MovementAmount'
        reference:
          type: string
          description: Free-text reference carried on the movement.
          maxLength: 140
          example: salary
    Account:
      type: object
      required: [accountUid, name, currency, createdAt, owner]
      properties:
        accountUid:
          type: string
          format: uuid
          example: f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50
        name:
          type: string
          example: ACC-001
        currency:
          type: string
          description: ISO-4217 alphabetic code.
          pattern: '^[A-Z]{3}$'
          example: GBP
        createdAt:
          type: string
          format: date-time
          example: '2026-08-09T09:15:00Z'
        owner:
          type: string
          description: The principal that opened the account — the JWT subject.
          example: alice
    AccountList:
      type: object
      description: Wrapped in a named key so siblings can be added without breaking clients (§7).
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
          example:
            - accountUid: f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50
              name: ACC-001
              currency: GBP
              createdAt: '2026-08-09T09:15:00Z'
              owner: alice
...
    Transaction:
      type: object
      description: |
        `direction` carries the sign so `amount` is always positive; `type` is the orthogonal axis
        saying what kind of movement it was (§7).
      required:
        - transactionUid
        - accountUid
        - type
        - direction
        - amount
        - balanceAfter
        - status
        - transactionTime
        - settlementTime
      properties:
        transactionUid:
          type: string
          format: uuid
          description: For a movement, the client-generated UID it was written to (§6.3).
          example: 1f0a9d3c-8e42-4a55-9b1e-6c7d8e9f0a1b
        accountUid:
          type: string
          format: uuid
          example: f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, ASSET_TRANSFER]
          example: DEPOSIT
        direction:
          type: string
          enum: [IN, OUT]
          example: IN
        amount:
          $ref: '#/components/schemas/Money'
        balanceAfter:
          $ref: '#/components/schemas/Money'
        status:
          type: string
          description: |
            Only `SETTLED` is produced today — appends settle atomically. `PENDING` and `REVERSED`
            are reserved so pending states never need a breaking change (§7).
          enum: [SETTLED, PENDING, REVERSED]
          example: SETTLED
        transactionTime:
          type: string
          format: date-time
          example: '2026-08-09T09:20:00Z'
        settlementTime:
          type: string
          format: date-time
          description: Equal to `transactionTime` today, for the same reason `status` is reserved.
          example: '2026-08-09T09:20:00Z'
        reference:
          type: string
          example: salary
...
    AssetTransferRequest:
      type: object
      required: [direction, assetSymbol, assetClass, quantity]
      properties:
        direction:
          type: string
          enum: [IN, OUT]
          example: IN
        assetSymbol:
          type: string
          minLength: 1
          maxLength: 10
          example: VOO
        assetClass:
          $ref: '#/components/schemas/AssetClassEnum'
        quantity:
          type: string
          description: Fractional share quantity with up to 6 decimal places.
          example: '10.500000'
        costBasis:
          $ref: '#/components/schemas/Money'
        lotId:
          type: string
          description: Optional acquisition lot ID (for IN direction).
          example: lot-voo-001
        selector:
          $ref: '#/components/schemas/TaxLotSelectorEnum'
        reference:
          type: string
          maxLength: 140
          example: VOO ETF allocation
    AssetTransaction:
      type: object
      required:
        - transactionUid
        - accountUid
        - type
        - direction
        - assetSymbol
        - assetClass
        - quantity
        - costBasis
        - status
        - transactionTime
        - settlementTime
      properties:
        transactionUid:
          type: string
          format: uuid
          example: 3b2a19c7-4e82-4f06-b9d3-8e7d5c4b1a20
        accountUid:
          type: string
          format: uuid
          example: f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, ASSET_TRANSFER]
          example: ASSET_TRANSFER
        direction:
          type: string
          enum: [IN, OUT]
          example: IN
        assetSymbol:
          type: string
          example: VOO
        assetClass:
          $ref: '#/components/schemas/AssetClassEnum'
        quantity:
          type: string
          example: '10.500000'
        costBasis:
          $ref: '#/components/schemas/Money'
        taxLots:
          type: array
          items:
            $ref: '#/components/schemas/TaxLotDto'
        balanceAfter:
          $ref: '#/components/schemas/Money'
        status:
          type: string
          enum: [SETTLED, PENDING, REVERSED]
          example: SETTLED
        transactionTime:
          type: string
          format: date-time
          example: '2026-08-21T10:00:00Z'
        settlementTime:
          type: string
          format: date-time
          example: '2026-08-21T10:00:00Z'
        reference:
          type: string
          example: VOO ETF allocation
    ProblemDetail:
      type: object
      description: |
        RFC 7807 problem detail (§6.5). No stack traces, no internal identifiers, no SQL fragments
        cross the boundary.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: The machine-readable error code — the catalogue in §6.5.
          example: '/errors/insufficient-funds'
        title:
          type: string
          example: Insufficient funds
        status:
          type: integer
          format: int32
          example: 422
        detail:
          type: string
          example: The account balance is lower than the requested withdrawal.
        instance:
          type: string
          format: uri-reference
          example: '/api/v1/accounts/f91e6c0e-1f3d-4b2a-9c77-0b1c2d3e4f50/withdrawals/2c1b8e47-3d95-4f06-a8b2-7e5d4c3b2a19'
        traceId:
          type: string
          description: Correlates the failure to the tracing backend (§6.6).
          example: 7c6f2e1a9b4d5c30